Introduce a uniform formatting style
I think I found a set of settings that doesn't cause a plethora of
changes and provides uniformity. Unfortunately, there is a mix of
different styles all across the codebase, both among actual minidlna
files and files taken from miniupnp. Hopefully this makes all this text
a tad easier to the eye...